Refactor logging and app structure
1. Removing 'monad-logger'

The first change was to remove 'monad-logger' and instead roll a custom typeclass. Why? The former necessitates using 'LoggingT' in our stack, which would normally be fine, but here it is awkward. Some commands like 'find-stale' (and forthcoming CLI actions) necessarily combine multiple 'IO' actions into one because of parallelism (resp. concurrency). In some circumstances, we would like logging in the individual 'IO' action, but of course we do not have access to 'LoggingT' in 'IO'. We could live with multiple 'levels' of logging, but that is a bit awkward in the presence of formatting (e.g. colors, prefixes like [Info]) since it requires one of:

  a. Re-implement 'LoggingT' style in 'IO'.

  b. Give up on having a consistent presentation.

There may be a way around this with 'unliftIO', but that is heavy-handed (and possibly naive given multithreading) when there is a much simpler solution: making a simple MonadLogger typeclass that we can implement in 'IO' (or a pure monad for testing) and then lift onto any transformer stack. We have opted for this simpler solution.

2. Moving typeclass instances from transformer to base monad

This refactor brought to light a previous design decision: defining the app's typeclass instances in terms of a concrete-ish transformer stack like 'ReaderT e m a'; we would therefore have a stack like 'MonadIO m => ReaderT e m a' for 'IO' and 'ReaderT e PureMonad a' for tests. 'ReaderT's ability to provide clean function signatures is seductive, but this is probably sub-optimal at the typeclass level because:

  a. Arguably the hidden parameters are fundamental to the tyepclasss actions, e.g., 'getStaleLogs' really should have params like the current day and filepath in its signature.

  b. More importantly, given the logging changes above, if we kept the status quo then we would have multiple levels of typeclass instances, at both the transformer and base monad levels. Perhaps this is unavoidable in larger applications, but in this case it is simpler to keep the instances in the base monad. This has an additional advantage that we can easily use a single transformer in both 'IO' and pure tests. All we need to do is change the base monad to get the desired behavior, so instead we end up with just 'ReaderT e IO a' and 'ReaderT e PureMonad a'.